+++ common/src/main/java/mca/Config.java
@@ -64,6 +64,8 @@ public final class Config implements Serializable {
     public boolean showNameTags = true;
     public int villagerHealthBonusPerLevel = 5;
     public boolean useSquidwardModels = false;
+    public int maxBuildingSize = 8192;
+    public int maxBuildingRadius = 320;
 
     public static File getConfigFile() {
         return new File("./config/mca.json");