+++ changelog.md
@@ -214,3 +214,4 @@
 * Converted villagers will now retain custom nbt data and age
 * Fixed inventory disappearing on convert
 * Fixed marriage and family tree loss on convert
+* Maximum building size and radius are now configurable